widgets core module compile success
