+++ widgets/src/iosMain/kotlin/dev/icerock/moko/widgets/core/screen/Screen.kt
@@ -7,8 +7,7 @@ package dev.icerock.moko.widgets.core.screen
 import dev.icerock.moko.graphics.Color
 import dev.icerock.moko.mvvm.dispatcher.EventsDispatcher
 import dev.icerock.moko.mvvm.viewmodel.ViewModel
-import dev.icerock.moko.widgets.core.objc.getAssociatedObject
-import dev.icerock.moko.widgets.core.objc.setAssociatedObject
+import dev.icerock.moko.widgets.core.associatedObject
 import platform.UIKit.UIViewController
 import kotlin.native.ref.WeakReference
 import kotlin.properties.ReadOnlyProperty
@@ -51,7 +50,7 @@ actual abstract class Screen<Arg : Args> {
             if (current != null) return current
 
             val vc = createViewController(isLightStatusBar).also {
-                setAssociatedObject(it, this)
+                it.associatedObject = this
             }
             onViewCreated()
             _viewController = WeakReference(vc)
@@ -67,4 +66,4 @@ actual abstract class Screen<Arg : Args> {
     }
 }
 
-fun UIViewController.getAssociatedScreen(): Screen<*>? = getAssociatedObject(this) as? Screen<*>
+fun UIViewController.getAssociatedScreen(): Screen<*>? = associatedObject as? Screen<*>
